requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
python-dotenv
psycopg2-binary
